Remove islandora_rest call
The call to islandora_rest works fine with varnish when you have a
short time frame between updating page number, but when you let the
page sit for a bit and then try to change the page number the call
to islandora_rest fails, preventing you from updating the page
number.  This commit just puts back the original pre-islandora_rest
way of updating the tile_source.